+++ psycopg/psycopg/cursor_async.py
@@ -164,7 +164,7 @@ class AsyncCursor(BaseCursor["AsyncConnection[Any]", Row]):
                 if self._pgconn.transaction_status == ACTIVE:
                     # Try to cancel the query, then consume the results
                     # already received.
-                    self._conn.cancel()
+                    await self._conn.cancel_safe()
                     try:
                         while await self._conn.wait(
                             self._stream_fetchone_gen(first=False)