+++ tests/test_generators.py
@@ -1,3 +1,4 @@
+import time
 from collections import deque
 from functools import partial
 from typing import List
@@ -42,6 +43,30 @@ def test_connect_operationalerror_pgconn(generators, dsn, monkeypatch):
     assert pgconn.pipeline_status == pq.PipelineStatus.OFF.value
     with pytest.raises(psycopg.OperationalError, match="connection is closed"):
         pgconn.exec_(b"select 1")
+
+
+@pytest.mark.libpq(">= 17")
+def test_cancel(pgconn, conn, generators):
+    pgconn.send_query_params(b"SELECT pg_sleep($1)", [b"180"])
+    while not conn.execute(
+        "SELECT count(*) FROM pg_stat_activity"
+        " WHERE query = 'SELECT pg_sleep($1)'"
+        " AND state = 'active'"
+    ).fetchone():
+        time.sleep(0.01)
+    cancel_conn = pgconn.cancel_conn()
+    assert cancel_conn.status != pq.ConnStatus.BAD
+    cancel_conn.start()
+    gen = generators.cancel(cancel_conn)
+    waiting.wait_conn(gen)
+    assert cancel_conn.status == pq.ConnStatus.OK
+
+    res = pgconn.get_result()
+    assert res is not None
+    assert res.status == pq.ExecStatus.FATAL_ERROR
+    assert res.error_field(pq.DiagnosticField.SQLSTATE) == b"57014"
+    while pgconn.is_busy():
+        pgconn.consume_input()
 
 
 @pytest.fixture